feat: refactor theme handling to update native theme asynchronously in ThemeProvider
